improved nonce factory
