+++ xchange-bitbay/src/main/java/org/knowm/xchange/bitbay/v3/BitbayExchange.java
@@ -5,8 +5,10 @@ import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeSpecification;
 import org.knowm.xchange.bitbay.v3.service.BitbayAccountService;
 import org.knowm.xchange.bitbay.v3.service.BitbayTradeService;
-import org.knowm.xchange.utils.nonce.CurrentTime1000NonceFactory;
+import org.knowm.xchange.utils.nonce.CurrentTimeIncrementalNonceFactory;
 import si.mazi.rescu.SynchronizedValueFactory;
+
+import java.util.concurrent.TimeUnit;
 
 /**
  * Initial support for the new BitBay API. Official documentation isn't even out yet so use this at
@@ -17,7 +19,8 @@ import si.mazi.rescu.SynchronizedValueFactory;
  */
 public class BitbayExchange extends BaseExchange implements Exchange {
 
-  private SynchronizedValueFactory<Long> nonceFactory = new CurrentTime1000NonceFactory();
+  private final SynchronizedValueFactory<Long> nonceFactory =
+          new CurrentTimeIncrementalNonceFactory(TimeUnit.SECONDS);
 
   @Override
   public ExchangeSpecification getDefaultExchangeSpecification() {
