--- xchange-core/src/main/java/org/knowm/xchange/utils/nonce/CurrentTime1000NonceFactory.java
@@ -1,12 +1,0 @@
-package org.knowm.xchange.utils.nonce;
-
-import si.mazi.rescu.SynchronizedValueFactory;
-
-public class CurrentTime1000NonceFactory implements SynchronizedValueFactory<Long> {
-
-  @Override
-  public Long createValue() {
-
-    return System.currentTimeMillis() / 1000L;
-  }
-}
