+++ xchange-exmo/src/main/java/org/knowm/xchange/exmo/ExmoExchange.java
@@ -2,6 +2,8 @@ package org.knowm.xchange.exmo;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.concurrent.TimeUnit;
+
 import org.knowm.xchange.BaseExchange;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeSpecification;
@@ -10,12 +12,12 @@ import org.knowm.xchange.exceptions.ExchangeException;
 import org.knowm.xchange.exmo.service.ExmoAccountService;
 import org.knowm.xchange.exmo.service.ExmoMarketDataService;
 import org.knowm.xchange.exmo.service.ExmoTradeService;
-import org.knowm.xchange.utils.nonce.CurrentNanoIncrementalNonceFactory;
+import org.knowm.xchange.utils.nonce.CurrentTimeIncrementalNonceFactory;
 import si.mazi.rescu.SynchronizedValueFactory;
 
 public class ExmoExchange extends BaseExchange implements Exchange {
   private final SynchronizedValueFactory<Long> nonceFactory =
-      new CurrentNanoIncrementalNonceFactory();
+      new CurrentTimeIncrementalNonceFactory(TimeUnit.NANOSECONDS);
 
   @Override
   protected void initServices() {
